Tokens/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
bitarray==2.7.2
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.12.1
Django==4.1.7
djangorestframework==3.14.0
drf-yasg==1.21.5
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
inflection==0.5.1
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.2
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
packaging==23.0
parsimonious==0.8.1
protobuf==3.19.5
pycryptodome==3.17
pyrsistent==0.19.3
python-decouple==3.7
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
rlp==2.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
toolz==0.12.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
varint==1.0.2
web3==5.31.3
websockets==9.1
yarl==1.8.2
